extracted interfaces for http and queue. Will be possible to mock and tests with these.
